構造を変更 責務を分解
